created new modules unbundling cli and updated cli_usage.md
