+++ abidex/workflows/__init__.py
@@ -1,0 +1,18 @@
+from .commands import (
+    list_workflows,
+    show_workflow_logs,
+    show_workflow_map,
+    open_workflow_notebook,
+)
+from .registry import WorkflowDescription, WorkflowRegistry
+from .discovery import discover_workflows
+
+__all__ = [
+    "WorkflowDescription",
+    "WorkflowRegistry",
+    "discover_workflows",
+    "list_workflows",
+    "show_workflow_logs",
+    "show_workflow_map",
+    "open_workflow_notebook",
+]
